move definition of TypedDicts into separate file
