+++ src/libkernelbot/leaderboard_db.py
@@ -1,10 +1,11 @@
 import dataclasses
 import datetime
 import json
-from typing import Dict, List, NotRequired, Optional, TypedDict
+from typing import Dict, List, Optional
 
 import psycopg2
 
+from libkernelbot.db_types import LeaderboardItem, LeaderboardRankedEntry, RunItem, SubmissionItem
 from libkernelbot.run_eval import CompileResult, RunResult, SystemInfo
 from libkernelbot.task import LeaderboardDefinition, LeaderboardTask
 from libkernelbot.utils import (
@@ -985,56 +986,6 @@ class LeaderboardDB:
             self.connection.rollback()
             logger.exception("Error validating CLI ID %s", cli_id, exc_info=e)
             raise KernelBotError("Error validating CLI ID") from e
-
-
-class LeaderboardItem(TypedDict):
-    id: int
-    name: str
-    creator_id: int
-    deadline: datetime.datetime
-    description: str
-    task: "LeaderboardTask"
-    gpu_types: List[str]
-    forum_id: int
-    secret_seed: NotRequired[int]
-
-
-class LeaderboardRankedEntry(TypedDict):
-    submission_id: int
-    rank: int
-    submission_name: str
-    submission_time: datetime.datetime
-    submission_score: float
-    leaderboard_name: str
-    user_id: int
-    user_name: str
-    gpu_type: str
-
-
-class RunItem(TypedDict):
-    start_time: datetime.datetime
-    end_time: datetime.datetime
-    mode: str
-    secret: bool
-    runner: str
-    score: Optional[float]
-    passed: bool
-    compilation: dict
-    meta: dict
-    result: dict
-    system: dict
-
-
-class SubmissionItem(TypedDict):
-    submission_id: int
-    leaderboard_id: int
-    leaderboard_name: str
-    file_name: str
-    user_id: int
-    submission_time: datetime.datetime
-    done: bool
-    code: str
-    runs: List[RunItem]
 
 
 class LeaderboardDoesNotExist(KernelBotError):
